backup/requirements.txt
# Web アプリケーションフレームワーク
streamlit==1.50.0

# LangChain関連
langchain==0.3.27
langchain-openai==0.3.34

# OpenAI API
openai==2.1.0

# 環境変数管理
python-dotenv==1.1.1

# データ処理（オプション）
pandas==2.3.3
numpy==2.3.3